Fix bug in lookup_transition()
- lookup_transition() was looking for a nullptr end marker in the table
array, but since the array doesn't have a nullptr end marker but
actually a valid structure with nullptr members as an end marker, the
function will never terminate if presented with a original state
function pointer that is not in the table. This fix corrects that
by comparing the structure members with nullptr to determine the end
of the table, and returning nullptr appropriately, when the function
is called to lookup the next state for a function pointer that is not
in the state table.